requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
x402[fastapi,httpx,evm]
eth-account==0.13.4
web3==7.6.0
google-genai==0.3.0
streamlit==1.41.1
python-dotenv==1.0.1
circle-developer-controlled-wallets
pydantic==2.10.4
